# src/specops_workshop/v4/generated/openapi/workshop-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactConfirmationReceipt": {
        "additionalProperties": false,
        "properties": {
          "binding": {
            "$ref": "#/components/schemas/ArtifactConfirmationBinding"
          },
          "command": {
            "$ref": "#/components/schemas/FoundationCommandReceipt"
          },
          "receipt_type": {
            "const": "ARTIFACT_CONFIRMATION",
            "title": "Receipt Type",
            "type": "string"
          }
        },
        "required": [
          "receipt_type",
          "command",
          "binding"
        ],
        "title": "ArtifactConfirmationReceipt",
        "type": "object"
      },
...
      "ArtifactReviewIntent": {
        "additionalProperties": false,
        "properties": {
          "operation_key": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,255}$",
            "title": "Operation Key",
            "type": "string"
          }
        },
        "required": [
          "operation_key"
        ],
        "title": "ArtifactReviewIntent",
        "type": "object"
      },
...
      "ArtifactReviewReceipt": {
        "additionalProperties": false,
        "properties": {
          "command": {
            "$ref": "#/components/schemas/FoundationCommandReceipt"
          },
          "confirmation_id": {
            "format": "uuid",
            "title": "Confirmation Id",
            "type": "string"
          },
          "receipt_type": {
            "const": "ARTIFACT_REVIEW",
            "title": "Receipt Type",
            "type": "string"
          },
          "view_hash": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "View Hash",
            "type": "string"
          },
          "view_id": {
            "format": "uuid",
            "title": "View Id",
            "type": "string"
          }
        },
        "required": [
          "receipt_type",
          "command",
          "confirmation_id",
          "view_id",
          "view_hash"
        ],
        "title": "ArtifactReviewReceipt",
        "type": "object"
      },
...
      "ArtifactSynthesisIntent": {
        "additionalProperties": false,
        "properties": {
          "operation_key": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,255}$",
            "title": "Operation Key",
            "type": "string"
          }
        },
        "required": [
          "operation_key"
        ],
        "title": "ArtifactSynthesisIntent",
        "type": "object"
      },
      "ArtifactSynthesisResult": {
        "additionalProperties": false,
        "properties": {
          "quality_audit": {
            "$ref": "#/components/schemas/ArtifactQualityAuditReceipt"
          },
          "synthesis": {
            "$ref": "#/components/schemas/ArtifactSynthesisAdmissionReceipt"
          }
        },
        "required": [
          "synthesis",
          "quality_audit"
        ],
        "title": "ArtifactSynthesisResult",
        "type": "object"
      },
...
      "ContractValidationEnvelope": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ContractValidationResponse"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ContractValidationEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "V4_EnterpriseSsoAuthentication": {
        "additionalProperties": false,
        "properties": {
          "actor_id": {
            "format": "uuid",
            "title": "Actor Id",
            "type": "string"
          },
          "assurance_level": {
            "const": "VERIFIED",
            "title": "Assurance Level",
            "type": "string"
          },
          "authentication_event_id": {
            "format": "uuid",
            "title": "Authentication Event Id",
            "type": "string"
          },
          "authentication_method": {
            "const": "ENTERPRISE_SSO",
            "title": "Authentication Method",
            "type": "string"
          },
          "external_subject": {
            "maxLength": 255,
            "minLength": 1,
            "title": "External Subject",
            "type": "string"
          },
          "identity_provider": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Identity Provider",
            "type": "string"
          }
        },
        "required": [
          "authentication_method",
          "assurance_level",
          "actor_id",
          "identity_provider",
          "external_subject",
          "authentication_event_id"
        ],
        "title": "EnterpriseSsoAuthentication",
        "type": "object"
      },
...
      "V4_VerbalSelfAssertion": {
        "additionalProperties": false,
        "properties": {
          "actor_id": {
            "format": "uuid",
            "title": "Actor Id",
            "type": "string"
          },
          "asserted_display_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Asserted Display Name",
            "type": "string"
          },
          "assertion_transcript_event_id": {
            "format": "uuid",
            "title": "Assertion Transcript Event Id",
            "type": "string"
          },
          "assurance_level": {
            "const": "SELF_ASSERTED",
            "title": "Assurance Level",
            "type": "string"
          },
          "authentication_method": {
            "const": "VERBAL_SELF_ASSERTION",
            "title": "Authentication Method",
            "type": "string"
          },
          "claimed_role": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Claimed Role",
            "type": "string"
          }
        },
        "required": [
          "authentication_method",
          "assurance_level",
          "actor_id",
          "asserted_display_name",
          "claimed_role",
          "assertion_transcript_event_id"
        ],
        "title": "VerbalSelfAssertion",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SpecOps Workshop Protocol",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v4/artifacts/current/confirm": {
      "post": {
        "operationId": "confirm_current_artifact_api_v4_artifacts_current_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "actor_authentication": {
                    "discriminator": {
                      "mapping": {
                        "ENTERPRISE_SSO": "#/components/schemas/V4_EnterpriseSsoAuthentication",
                        "VERBAL_SELF_ASSERTION": "#/components/schemas/V4_VerbalSelfAssertion"
                      },
                      "propertyName": "authentication_method"
                    },
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/V4_VerbalSelfAssertion"
                      },
                      {
                        "$ref": "#/components/schemas/V4_EnterpriseSsoAuthentication"
                      }
                    ],
                    "title": "Actor Authentication"
                  },
                  "confirmation_transcript_event_id": {
                    "format": "uuid",
                    "title": "Confirmation Transcript Event Id",
                    "type": "string"
                  }
                },
                "required": [
                  "actor_authentication",
                  "confirmation_transcript_event_id"
                ],
                "title": "ArtifactConfirmationIntent",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactConfirmationReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractValidationEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Confirm Current Artifact",
        "tags": [
          "Workshop Protocol 1.0.0"
        ]
      }
    },
    "/api/v4/artifacts/{artifact_type}/review": {
      "post": {
        "operationId": "materialize_artifact_review_api_v4_artifacts__artifact_type__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_type",
            "required": true,
            "schema": {
              "pattern": "^(SPEC_PACKAGE|TECHNICAL_CONTRACT)$",
              "title": "Artifact Type",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactReviewIntent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactReviewReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Materialize Artifact Review",
        "tags": [
          "Workshop Protocol 1.0.0"
        ]
      }
    },
    "/api/v4/artifacts/{artifact_type}/synthesize": {
      "post": {
        "operationId": "synthesize_artifact_api_v4_artifacts__artifact_type__synthesize_post",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_type",
            "required": true,
            "schema": {
              "pattern": "^(SPEC_PACKAGE|TECHNICAL_CONTRACT)$",
              "title": "Artifact Type",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactSynthesisIntent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactSynthesisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Synthesize Artifact",
        "tags": [
          "Workshop Protocol 1.0.0"
        ]
      }
    },
    "/api/v4/cases/{case_id}/artifact-review": {
      "get": {
        "operationId": "current_artifact_review_api_v4_cases__case_id__artifact_review_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ArtifactReviewProjection"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Current Artifact Review Api V4 Cases  Case Id  Artifact Review Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Current Artifact Review",
        "tags": [
          "Workshop Protocol 1.0.0"
        ]
      }
    },
    "/api/v4/cases/{case_id}/artifacts/{artifact_type}": {
      "get": {
        "operationId": "confirmed_artifact_api_v4_cases__case_id__artifacts__artifact_type__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_type",
            "required": true,
            "schema": {
              "pattern": "^(SPEC_PACKAGE|TECHNICAL_CONTRACT)$",
              "title": "Artifact Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConfirmedArtifactProjection"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Confirmed Artifact Api V4 Cases  Case Id  Artifacts  Artifact Type  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirmed Artifact",
        "tags": [
          "Workshop Protocol 1.0.0"
        ]
      }
    },
    "/api/v4/cases/{case_id}/decision-review": {
      "get": {
        "operationId": "current_decision_review_api_v4_cases__case_id__decision_review_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DecisionBatchReviewView"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Current Decision Review Api V4 Cases  Case Id  Decision Review Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Current Decision Review",
        "tags": [
          "Workshop Protocol 1.0.0"
        ]
      }
    },
    "/api/v4/decisions/current/respond": {
      "post": {
        "operationId": "apply_browser_decision_selection_api_v4_decisions_current_respond_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "actor_authentication": {
                    "discriminator": {
                      "mapping": {
                        "ENTERPRISE_SSO": "#/components/schemas/V4_EnterpriseSsoAuthentication",
                        "VERBAL_SELF_ASSERTION": "#/components/schemas/V4_VerbalSelfAssertion"
                      },
                      "propertyName": "authentication_method"
                    },
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/V4_VerbalSelfAssertion"
                      },
                      {
                        "$ref": "#/components/schemas/V4_EnterpriseSsoAuthentication"
                      }
                    ],
                    "title": "Actor Authentication"
                  },
                  "operation_key": {
                    "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,255}$",
                    "title": "Operation Key",
                    "type": "string"
                  },
                  "response_transcript_event_id": {
                    "format": "uuid",
                    "title": "Response Transcript Event Id",
                    "type": "string"
                  },
                  "selections": {
                    "items": {
                      "$ref": "#/components/schemas/V4_VoiceConfirmationSelectionItemCandidate"
                    },
                    "maxItems": 26,
                    "minItems": 1,
                    "title": "Selections",
                    "type": "array"
                  }
                },
                "required": [
                  "operation_key",
                  "response_transcript_event_id",
                  "actor_authentication",
                  "selections"
                ],
                "title": "DecisionResponseIntent",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionBatchResponseReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractValidationEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Apply Browser Decision Selection",
        "tags": [
          "Workshop Protocol 1.0.0"
        ]
      }
    },
...
    "/api/v4/voice/decision-selections": {
      "post": {
        "operationId": "apply_voice_decision_selection_api_v4_voice_decision_selections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "actor_authentication": {
                    "discriminator": {
                      "mapping": {
                        "ENTERPRISE_SSO": "#/components/schemas/V4_EnterpriseSsoAuthentication",
                        "VERBAL_SELF_ASSERTION": "#/components/schemas/V4_VerbalSelfAssertion"
                      },
                      "propertyName": "authentication_method"
                    },
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/V4_VerbalSelfAssertion"
                      },
                      {
                        "$ref": "#/components/schemas/V4_EnterpriseSsoAuthentication"
                      }
                    ],
                    "title": "Actor Authentication"
                  },
                  "selection": {
                    "$ref": "#/components/schemas/V4_VoiceConfirmationSelectionCandidate"
                  }
                },
                "required": [
                  "selection",
                  "actor_authentication"
                ],
                "title": "VoiceSelectionInput",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionBatchResponseReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractValidationEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Apply Voice Decision Selection",
        "tags": [
          "Workshop Protocol 1.0.0"
        ]
      }
    },
...
  }
}